wcnss: add sysfs entry to support Thermal Mitigation
The Thermal Mitigation function in userspace needs a mechanism by
which it can notify the WCNSS driver that thermal mitigation is
required.  We introduce a sysfs entry via which Thermal Mitigation can
both read the current mitigation level as well as set a new mitigation
level.  Logic to actually act upon a new mitigation level will be
added at a later time.

Change-Id: I1b2b8328e2aa821414de0d345bd32b9bc30bd49a
Signed-off-by: Jeff Johnson <jjohnson@codeaurora.org>

Conflicts:

	drivers/net/wireless/wcnss/wcnss_wlan.c